requirements.txt
numpy
torch
matplotlib
pyyaml
pygame
random
os
time